requirements.txt
beautifulsoup4==4.9.3
selenium==3.141.0
requests==2.25.1
python-telegram-bot==13.3
